test(build): consolidate compatibility coverage under postbuild
